+++ CHANGELOG.md
@@ -79,6 +79,8 @@ Main (unreleased)
 
 - Fix `pyroscope.receive_http` so it does not restart server if the server configuration has not changed. (@korniltsev)
 
+- Fix default connection limit in `pyroscope.receive_http`. (@korniltsev)
+
 v1.10.1
 -----------------